Unify RoleDescriptor, add security.update_api_key API
