+++ output/typescript/types.ts
@@ -15218,6 +15218,28 @@ export interface SecurityRealmInfo {
   type: string
 }
 
+export interface SecurityRoleDescriptor {
+  cluster?: string[]
+  indices?: SecurityIndicesPrivileges[]
+  index?: SecurityIndicesPrivileges[]
+  global?: SecurityGlobalPrivilege[] | SecurityGlobalPrivilege
+  applications?: SecurityApplicationPrivileges[]
+  metadata?: Metadata
+  run_as?: string[]
+  transient_metadata?: SecurityTransientMetadataConfig
+}
+
+export interface SecurityRoleDescriptorRead {
+  cluster: string[]
+  indices: SecurityIndicesPrivileges[]
+  index: SecurityIndicesPrivileges[]
+  global?: SecurityGlobalPrivilege[] | SecurityGlobalPrivilege
+  applications?: SecurityApplicationPrivileges[]
+  metadata?: Metadata
+  run_as?: string[]
+  transient_metadata?: SecurityTransientMetadataConfig
+}
+
 export interface SecurityRoleMapping {
   enabled: boolean
   metadata: Metadata
@@ -15390,7 +15412,7 @@ export interface SecurityCreateApiKeyRequest extends RequestBase {
   body?: {
     expiration?: Duration
     name?: Name
-    role_descriptors?: Record<string, SecurityCreateApiKeyRoleDescriptor>
+    role_descriptors?: Record<string, SecurityRoleDescriptor>
     metadata?: Metadata
   }
 }
@@ -15401,17 +15423,6 @@ export interface SecurityCreateApiKeyResponse {
   id: Id
   name: Name
   encoded: string
-}
-
-export interface SecurityCreateApiKeyRoleDescriptor {
-  cluster: string[]
-  indices: SecurityIndicesPrivileges[]
-  index: SecurityIndicesPrivileges[]
-  global?: SecurityGlobalPrivilege[] | SecurityGlobalPrivilege
-  applications?: SecurityApplicationPrivileges[]
-  metadata?: Metadata
-  run_as?: string[]
-  transient_metadata?: SecurityTransientMetadataConfig
 }
 
 export interface SecurityCreateServiceTokenRequest extends RequestBase {
@@ -15600,19 +15611,8 @@ export interface SecurityGetServiceAccountsRequest extends RequestBase {
 
 export type SecurityGetServiceAccountsResponse = Record<string, SecurityGetServiceAccountsRoleDescriptorWrapper>
 
-export interface SecurityGetServiceAccountsRoleDescriptor {
-  cluster: string[]
-  indices: SecurityIndicesPrivileges[]
-  index: SecurityIndicesPrivileges[]
-  global?: SecurityGlobalPrivilege[] | SecurityGlobalPrivilege
-  applications?: SecurityApplicationPrivileges[]
-  metadata?: Metadata
-  run_as?: string[]
-  transient_metadata?: SecurityTransientMetadataConfig
-}
-
 export interface SecurityGetServiceAccountsRoleDescriptorWrapper {
-  role_descriptor: SecurityGetServiceAccountsRoleDescriptor
+  role_descriptor: SecurityRoleDescriptorRead
 }
 
 export interface SecurityGetServiceCredentialsNodesCredentials {
@@ -15997,6 +15997,18 @@ export interface SecuritySuggestUserProfilesResponse {
 export interface SecuritySuggestUserProfilesTotalUserProfiles {
   value: long
   relation: RelationName
+}
+
+export interface SecurityUpdateApiKeyRequest extends RequestBase {
+  id: Id
+  body?: {
+    role_descriptors?: Record<string, SecurityRoleDescriptor>
+    metadata?: Metadata
+  }
+}
+
+export interface SecurityUpdateApiKeyResponse {
+  updated: boolean
 }
 
 export interface SecurityUpdateUserProfileDataRequest extends RequestBase {
