+++ specification/security/get_service_accounts/types.ts
@@ -17,25 +17,8 @@
  * under the License.
  */
 
-import {
-  IndicesPrivileges,
-  GlobalPrivilege,
-  ApplicationPrivileges
-} from '@security/_types/Privileges'
-import { TransientMetadataConfig } from '@security/_types/TransientMetadataConfig'
-import { Metadata } from '@_types/common'
+import { RoleDescriptorRead } from '@security/_types/RoleDescriptor'
 
 export class RoleDescriptorWrapper {
-  role_descriptor: RoleDescriptor
-}
-
-export class RoleDescriptor {
-  cluster: string[]
-  /** @aliases index */
-  indices: IndicesPrivileges[]
-  global?: GlobalPrivilege[] | GlobalPrivilege
-  applications?: ApplicationPrivileges[]
-  metadata?: Metadata
-  run_as?: string[]
-  transient_metadata?: TransientMetadataConfig
+  role_descriptor: RoleDescriptorRead
 }
